Add java test
1. add java test
2. change unsigned to signed because Java only support signed

+++ cczoo/cross_lang_framework/clf_client/app/test.c
@@ -8,26 +8,26 @@ int main()
 {
 	printf("test IN\n");
 	//secret_prov_test();
-	uint8_t key[64] = {0};
+	int8_t key[64] = {0};
 	get_key(key, 64);
 	printf("key=%s\n", key);
 
 	char *fname = "README.md";
-	uint64_t len = 0;
+	int64_t len = 0;
 	get_file_size(fname, &len);
 	printf("len=%lu\n", len);
 
-	uint8_t* data = (uint8_t*)malloc(len);
+	int8_t* data = (int8_t*)malloc(len);
 	if(!data) {
 		printf("malloc(%lu) fail\n", len);
 		return -1;
 	}
 	memset(data, 0, len);
-	uint64_t ret_len = 0;
+	int32_t ret_len = 0;
 	get_file_2_buff("README.md", 2, data, 10, &ret_len);
-	printf("read_len=%ld, data=%s\n", ret_len, data);
+	printf("read_len=%d, data=%s\n", ret_len, data);
 	get_file_2_buff("README.md", 0, data, 20, &ret_len);
-	printf("read_len=%ld, data=%s\n", ret_len, data);
+	printf("read_len=%d, data=%s\n", ret_len, data);
 
 	put_result("1.dat", 0, data, 20);
