+++ cczoo/cross_lang_framework/clf_client/clf_client.h
@@ -7,10 +7,11 @@
 extern "C" {
 #endif
 
-int get_key(uint8_t* key, uint32_t key_len);
-int get_file_2_buff(char* fname, uint64_t offset, uint8_t* data, uint64_t len, uint64_t* ret_len);
-int get_file_size(char* fname, uint64_t* ret_len);
-int put_result(char* fname, uint64_t offset, uint8_t* data, int32_t len);
+int get_key(int8_t* key, int32_t key_len);
+/* Java maximum array size is int32 */
+int get_file_2_buff(char* fname, int64_t offset, int8_t* data, int32_t len, int32_t* ret_len);
+int get_file_size(char* fname, int64_t* ret_len);
+int put_result(char* fname, int64_t offset, int8_t* data, int32_t len);
 
 #ifdef __cplusplus
 }
